// src/state/slices/opportunitiesSlice/types.ts
export enum DefiType {
  LiquidityPool = 'lp',
  Staking = 'staking',
}

export enum DefiProvider {
  Idle = 'Idle',
  ShapeShift = 'ShapeShift',
  EthJinxStaking = 'ETH/JINX Staking',
  UniV2 = 'Uniswap V2',
  CosmosSdk = 'Cosmos SDK',
  OsmosisLp = 'Osmosis LP',
  ThorchainSavers = 'THORChain Savers',
}

export type DefiProviderMetadata = {
  provider: string
  icon: string
  color: string
  url?: string
}

export type AssetIdsTuple =
  | readonly [AssetId, AssetId, AssetId]
  | readonly [AssetId, AssetId]
  | readonly [AssetId]
  | readonly []

export type OpportunityMetadataBase = {
  apy: string
  assetId: AssetId
  id: OpportunityId
  provider: string
  tvl: string
  type: DefiType
  // An optional user-facing `type` equivalent to allow us to display the opportunity type in the UI
  // our `type` property is an implementation detail that somehow ended up being user-facing, and might go away
  // This is actually more granular than the `type` property, i.e different opportunities of the same type/version might be part of different groups
  // Currently for read-only opportunities only
  group?: string
  // An opportunity might have its own icon e.g Cosmos SDK validators each have their own icon
  // If not specified, the underlying asset IDs' icons are used as icons
  icon?: string
  // For LP opportunities, this is the same as the AssetId
  // For staking opportunities i.e when you stake your LP asset, this is the AssetId of the LP asset being staked
  // Which might or might not be the same as the AssetId, e.g
  // - with LP staking, you stake an LP token
  // - with Idle, you stake a yield-bearing token, so the underlyingAssetId *is* the AssetId
  // - with Cosmos SDK, you stake the underlying asset directly, so the underlyingAssetId *is* the AssetId
  underlyingAssetId: AssetId
  // The AssetId or AssetIds this opportunity represents
  // For LP tokens, that's an asset pair
  // For opportunities a la Idle, that's the asset the opportunity wraps
  underlyingAssetIds: AssetIdsTuple
  // The underlying amount of underlyingAssetId 0 and maybe 1 per 1 LP token, in base unit
  underlyingAssetRatiosBaseUnit: readonly string[]
  // The reward assets this opportunity yields, typically 1/2 or 3 assets max.
  // Can also be empty in case there are no denominated rewards or we are unable to track them
  rewardAssetIds: AssetIdsTuple
  isClaimableRewards: boolean
  isReadOnly?: boolean
  // claimableRewards: boolean
  expired?: boolean
  active?: boolean
  name: string
  version?: string
  tags?: string[]
}

export type OpportunityMetadata =
  | OpportunityMetadataBase
  | ThorchainSaversStakingSpecificMetadata
  | IdleStakingSpecificMetadata

// User-specific values for this opportunity
export type UserStakingOpportunityBase = {
  userStakingId: UserStakingId
  // The amount of farmed LP tokens
  stakedAmountCryptoBaseUnit: string
  // The amount of rewards available to claim for the farmed LP position
  rewardsCryptoBaseUnit: {
    amounts:
      | readonly [string, string, string]
      | readonly [string, string]
      | readonly [string]
      | readonly []
    claimable: boolean
  }
}

export type UserStakingOpportunity =
  | UserStakingOpportunityBase
  | CosmosSdkStakingSpecificUserStakingOpportunity
  | JinxySpecificUserStakingOpportunity

export type UserStakingOpportunityWithMetadata = UserStakingOpportunity & OpportunityMetadata

// The AccountId of the staking contract in the form of chainId:accountAddress
export type StakingIdBase = Nominal<string, 'StakingId'> & AssetId
// Sometimes, an AssetIdish (i.e a a CAIP-19) is not enough to uniquely identify a staking opportunity
// So we add a StakingKey to the StakingId to make it unique
export type StakingKey = Nominal<string, 'StakingKey'>
export type StakingId = StakingIdBase | `${StakingIdBase}#${StakingKey}`
// The AccountId of the LP contract in the form of chainId:accountAddress
export type LpId = Nominal<string, 'LpId'> & AssetId

export type ValidatorId = Nominal<string, 'ValidatorId'> & AccountId
export type OpportunityId = LpId | StakingId | ValidatorId
// The unique identifier of an lp opportunity in the form of UserAccountId*StakingId
export type UserStakingId = `${AccountId}*${StakingId}`

export type OpportunitiesState = {
  lp: {
    byAccountId: PartialRecord<AccountId, LpId[]> // a 1:n foreign key of which user AccountIds hold this LpId
    byId: PartialRecord<LpId, OpportunityMetadataBase>
    ids: LpId[]
  }
  // Staking is the odd one here - it isn't a portfolio holding, but rather a synthetic value living on a smart contract
  // Which means we can't just derive the data from portfolio, marketData and other slices but have to actually store the amount in the slice
  userStaking: {
    byId: PartialRecord<UserStakingId, UserStakingOpportunity>
    ids: UserStakingId[]
  }

  staking: {
    // a 1:n foreign key of which user AccountIds hold this StakingId
    byAccountId: PartialRecord<AccountId, StakingId[]>
    byId: PartialRecord<StakingId, OpportunityMetadata>
    ids: StakingId[]
  }
}
// ...
export type GetOpportunityMetadataOutput = {
  byId: Record<OpportunityId, OpportunityMetadata>
  type: DefiType
}

export type GetReadOnlyOpportunitiesOutput = ReadOnlyOpportunityType[] | undefined

export type GetOpportunityUserDataOutput = {
  byAccountId: OpportunitiesState[DefiType]['byAccountId']
  type: DefiType
}

export type GetOpportunityUserStakingDataOutput = {
  byId: OpportunitiesState['userStaking']['byId']
}

export type GetOpportunityIdsOutput = OpportunityId[]